# html/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/public"
        }
    ],
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "Route to login",
                "operationId": "login",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "User's email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User's password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout (NECESSÁRIO ESTAR AUTENTICADO)",
                "description": "Rota para logout. Usuário DEVE estar autenticado para acessar.",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            }
        },
        "/api/v1/solicitar_recuperar_senha": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "solicitar recuperar senha",
                "description": "Route to solicitar_recuperar_senha",
                "operationId": "solicitar_recuperar_senha",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "User's email",
                                        "type": "string"
                                    },
                                    "link_definir_senha": {
                                        "description": "Link do front para definir nova senha",
                                        "type": "string url"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/v1/recuperar_senha": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Define nova senha para usuário",
                "description": "Define nova senha para usuário",
                "operationId": "recuperarSenha",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "description": "User's password",
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "description": "User's password confirmation",
                                        "type": "string"
                                    },
                                    "token": {
                                        "description": "Token de recuperação de senha. Enviado junto ao link no e-mail de recuperação.",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
        "/api/v1/user/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register user",
                "description": "Register the user after checking that they are a valid Colaborador",
                "operationId": "register",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "cpf": {
                                        "description": "Colaborador CPF",
                                        "type": "string"
                                    },
                                    "primeiro_nome": {
                                        "description": "Colaborador's first name",
                                        "type": "string"
                                    },
                                    "ultimo_nome": {
                                        "description": "Colaborador's last name",
                                        "type": "string"
                                    },
                                    "data_nascimento": {
                                        "description": "Colaborador's birthdate",
                                        "type": "date('d/m/Y')"
                                    },
                                    "email": {
                                        "description": "User's email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User's password",
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "description": "User's password confirmation",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "api_token": {
                "type": "apiKey",
                "description": "Adicione \"Bearer \" antes do token",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}